+++ src/lib.rs
@@ -234,8 +234,8 @@ pub use report::{
 pub use solver::{
     Dependencies, DependencyConstraints, DependencyProvider, IncompatibilityConstraint,
     IncompatibilityConstraintTerm, IncompatibilityConstraints, MaximalSolutions,
-    PackageResolutionStatistics, SelectedDependencies, resolve, resolve_maximal_solutions,
-    resolve_maximal_solutions_with_observer, resolve_with_observer,
+    PackageResolutionStatistics, SelectedDependencies, VersionOrdering, resolve,
+    resolve_maximal_solutions, resolve_maximal_solutions_with_observer, resolve_with_observer,
 };
 pub use term::Term;
 pub use type_aliases::{Map, Set};